updated directory links and moved logo to navbar folder
